+++ src/main/java/hu/ibello/model/ExpectationElement.java
@@ -15,9 +15,20 @@
  */
 package hu.ibello.model;
 
+import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name="expectation")
 public class ExpectationElement extends Element {
 
+	private long timeoutMs;
+	
+	public long getTimeoutMs() {
+		return timeoutMs;
+	}
+	
+	@XmlAttribute(name="timeout-ms")
+	public void setTimeoutMs(long timeoutMs) {
+		this.timeoutMs = timeoutMs;
+	}
 }